move api key to non-ee
