+++ backend/tests/integration/common_utils/managers/api_key.py
@@ -3,7 +3,7 @@ from uuid import uuid4
 import requests
 
 from danswer.db.models import UserRole
-from ee.danswer.server.api_key.models import APIKeyArgs
+from danswer.server.api_key.models import APIKeyArgs
 from tests.integration.common_utils.constants import API_SERVER_URL
 from tests.integration.common_utils.constants import GENERAL_HEADERS
 from tests.integration.common_utils.test_models import DATestAPIKey
